[CALCITE-2583] Upgrade dropwizard metrics to 4.0.3
