+++ metrics-dropwizardmetrics/src/main/resources/META-INF/services/org.apache.calcite.avatica.metrics.MetricsSystemFactory
@@ -1,0 +1,2 @@
+org.apache.calcite.avatica.metrics.dropwizard.DropwizardMetricsSystemFactory
+
