--- metrics-dropwizardmetrics3/src/main/java/org/apache/calcite/avatica/metrics/dropwizard3/DropwizardCounter.java
@@ -1,51 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to you under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.apache.calcite.avatica.metrics.dropwizard3;
-
-import com.codahale.metrics.Counter;
-
-import java.util.Objects;
-
-/**
- * Dropwizard Metrics implementation of {@link org.apache.calcite.avatica.metrics.Counter}.
- */
-public class DropwizardCounter implements org.apache.calcite.avatica.metrics.Counter {
-
-  private final Counter counter;
-
-  public DropwizardCounter(Counter counter) {
-    this.counter = Objects.requireNonNull(counter);
-  }
-
-  @Override public void increment() {
-    this.counter.inc();
-  }
-
-  @Override public void increment(long n) {
-    this.counter.inc(n);
-  }
-
-  @Override public void decrement() {
-    this.counter.dec();
-  }
-
-  @Override public void decrement(long n) {
-    this.counter.dec(n);
-  }
-}
-
-// End DropwizardCounter.java
