--- metrics-dropwizardmetrics3/src/main/java/org/apache/calcite/avatica/metrics/dropwizard3/DropwizardMetricsSystem.java
@@ -1,62 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to you under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.apache.calcite.avatica.metrics.dropwizard3;
-
-import org.apache.calcite.avatica.metrics.Counter;
-import org.apache.calcite.avatica.metrics.Gauge;
-import org.apache.calcite.avatica.metrics.Histogram;
-import org.apache.calcite.avatica.metrics.Meter;
-import org.apache.calcite.avatica.metrics.MetricsSystem;
-import org.apache.calcite.avatica.metrics.Timer;
-
-import com.codahale.metrics.MetricRegistry;
-
-import java.util.Objects;
-
-/**
- * Dropwizard Metrics implementation of {@link MetricsSystem}.
- */
-public class DropwizardMetricsSystem implements MetricsSystem {
-
-  private final MetricRegistry registry;
-
-  public DropwizardMetricsSystem(MetricRegistry registry) {
-    this.registry = Objects.requireNonNull(registry);
-  }
-
-  @Override public Timer getTimer(String name) {
-    return new DropwizardTimer(registry.timer(name));
-  }
-
-  @Override public Histogram getHistogram(String name) {
-    return new DropwizardHistogram(registry.histogram(name));
-  }
-
-  @Override public Meter getMeter(String name) {
-    return new DropwizardMeter(registry.meter(name));
-  }
-
-  @Override public Counter getCounter(String name) {
-    return new DropwizardCounter(registry.counter(name));
-  }
-
-  @Override public <T> void register(String name, Gauge<T> gauge) {
-    registry.register(name, new DropwizardGauge<T>(gauge));
-  }
-}
-
-// End DropwizardMetricsSystem.java
